OrderFragment.java - Switch to ConfirmPaidFragment and back to ProfileFragment => Completed
CartFragment.java
- Small change to totalPrice to show normal number instead of Exponential form

+++ app/src/main/java/com/example/e_commerce/network/model/response/order/Oder.java
@@ -3,10 +3,10 @@ package com.example.e_commerce.network.model.response.order;
 public class Oder {
     private String telephoneNumber;
     private String deliveryAddress;
-    private long statusPayment;
+    private int statusPayment;
     private long id;
     private String oderCode;
-    private long status;
+    private int status;
 
     public String getTelephoneNumber() { return telephoneNumber; }
     public void setTelephoneNumber(String value) { this.telephoneNumber = value; }
@@ -14,8 +14,8 @@ public class Oder {
     public String getDeliveryAddress() { return deliveryAddress; }
     public void setDeliveryAddress(String value) { this.deliveryAddress = value; }
 
-    public long getStatusPayment() { return statusPayment; }
-    public void setStatusPayment(long value) { this.statusPayment = value; }
+    public int getStatusPayment() { return statusPayment; }
+    public void setStatusPayment(int value) { this.statusPayment = value; }
 
     public long getid() { return id; }
     public void setid(long value) { this.id = value; }
@@ -23,6 +23,6 @@ public class Oder {
     public String getOderCode() { return oderCode; }
     public void setOderCode(String value) { this.oderCode = value; }
 
-    public long getStatus() { return status; }
-    public void setStatus(long value) { this.status = value; }
+    public int getStatus() { return status; }
+    public void setStatus(int value) { this.status = value; }
 }
